Fix Win32 test

+++ test/epochs.jl
@@ -1,5 +1,5 @@
 @testset "Epochs" begin
-    ep = Epoch2{TDB}(100000, 1e-18)
+    ep = Epoch2{TDB}(Int64(100000), 1e-18)
     ep1 = Epoch2{TDB}(ep, 100 * 365.25 * 86400)
     @test ep.offset == ep1.offset
 
@@ -10,7 +10,7 @@
     @test ep1.epoch == typemin(Int64)
     @test ep1.offset == -Inf
 
-    tai = Epoch2{TAI}(100000, 0.0)
+    tai = Epoch2{TAI}(Int64(100000), 0.0)
     tt = Epoch2{TT}(tai)
     @test tt.epoch == 100032
     @test tt.offset ≈ 0.184
